+++ richclient-project/.ubiquity/config.php
@@ -1,0 +1,2 @@
+<?php
+return array("host"=>"127.0.0.1","port"=>8090,"sessionName"=>"s5d1a47a6e2037");