+++ richclient-project/app/controllers/MaintenanceController.php
@@ -1,0 +1,10 @@
+<?php
+namespace controllers;
+
+use Ubiquity\controllers\admin\DefaultMaintenance;
+ /**
+ * Class MaintenanceController
+ **/
+class MaintenanceController extends DefaultMaintenance{
+
+}