+++ translations-project/app/translations/en_EN/messages.php
@@ -1,0 +1,2 @@
+<?php
+return array("bt.okay"=>"Okay","bt.cancel"=>"Cancel");