+++ CHANGELOG.md
@@ -8,3 +8,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - Auth-project : adds authentification to Admin controller 
+ - Translations-project : Demo for translations usage and creation
+ - Richclient-project : Demo of jQuery php functions and Semantic-UI components